COZMO-5863 Specialize I2C Bus (#2210)
* Trim

* Syncronous calls now work

* This is silly to not just replicate

* IMU is now the only read target, treat it as such

* Remove unnessessary constants

* Minor change

* IMU read moved so it can be used as a trigger instead

* IMU read state machine now functional

* Checkpoint

* Checkpoint

* Checkpoint

* Preliminary OLED pump

* Fixes

* Unjam seems to help alot

* Working I2C pump

* Return to old unjam style

* Use a bunch less ram

* Reasons

* Fix DAC feed crash, also improve performance

* Checkpoint

* Remove unjam, as it is no longer needed

* Working checkpoint

* Optimize code

* Minor optimization

* Checkpoint

* This does nothing

* Checkpoint

* Fixes

* Checkpoint

* Added watermarking

* serial image face tool

* Fix face corruption issue

* Better SPI communication loss compensation

* DEPS

+++ robot/k02_hal/hal/dac.cpp
@@ -87,7 +87,7 @@ void Anki::Cozmo::HAL::DAC::Sync() {
   // Set the write pointer to be the current read pointer minus one
   // The next feed will have moved ~MAX_AUDIO_BYTES_PER_DROP bytes 
   // forward, making this the ideal write location
-  write_pointer = ((DAC0_C2 >> 4) - 1) % DAC_WORDS;
+  write_pointer = (uint8_t)((DAC0_C2 >> 4) - 1) % DAC_WORDS;
 }
 
 static void adjustVolume(int target) {
